requirements.txt
flask==3.0.2
openai==1.3.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
gunicorn==21.2.0
